chore: refactor onto types.ts, and create custom errors
